des modification au niveau des templates

+++ src/elbaldi/GUI/ItemcategbackController.java
@@ -10,6 +10,7 @@ import elbaldi.services.CategorieCRUD;
 import java.io.IOException;
 import java.net.URL;
 import java.sql.SQLException;
+import java.util.Optional;
 import java.util.ResourceBundle;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
@@ -19,6 +20,8 @@ import javafx.scene.Parent;
 import javafx.scene.Scene;
 import javafx.scene.control.Alert;
 import javafx.scene.control.Button;
+import javafx.scene.control.ButtonBar;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.Label;
 import javafx.scene.layout.AnchorPane;
 import javafx.stage.Stage;
@@ -98,14 +101,30 @@ private int id_categorie;
     private void delete(ActionEvent event) throws SQLException, IOException {
        
       CategorieCRUD categoriecrud = new CategorieCRUD();
-      categoriecrud.supprimerCategorie(categorie1);
-  
-     
+      //categoriecrud.supprimerCategorie(categorie1);
+   Alert alert2=new Alert(Alert.AlertType.CONFIRMATION);
+       alert2.setTitle("Confirmation de suppression");
+       alert2.setHeaderText(null);
+       alert2.setContentText("Êtes-vous sûr de vouloir supprimer cette catégorie et les produits associes à cette categories ?");
+
+      ButtonType confirmerButton = new ButtonType("Confirmer", ButtonBar.ButtonData.OK_DONE);
+ButtonType cancelButton = new ButtonType("Annuler", ButtonBar.ButtonData.CANCEL_CLOSE);
+alert2.getButtonTypes().setAll(confirmerButton, cancelButton);
+
+Optional<ButtonType> result = alert2.showAndWait();
+if (result.get() == confirmerButton){
+      categoriecrud.supprimerCategorie(categorie1);       
       Alert alert = new Alert(Alert.AlertType.INFORMATION);
       alert.setTitle("Information Dialog");
       alert.setHeaderText(null);
-      alert.setContentText("categorie a été supprimé avec succés!");
+      alert.setContentText("catégorie a été supprimée avec succés!");
       alert.show();
+     
+//      Alert alert = new Alert(Alert.AlertType.INFORMATION);
+//      alert.setTitle("Information Dialog");
+//      alert.setHeaderText(null);
+//      alert.setContentText("categorie a été supprimé avec succés!");
+//      alert.show();
    Parent loader = FXMLLoader.load(getClass().getResource("catgeorielistBack.fxml"));
       nom_categfx.getScene().setRoot(loader);
       
@@ -122,4 +141,6 @@ private int id_categorie;
             
                 
         
-}}
+}
+    }
+}
